Remove Distributed Tuleap (distlp)
As the SVN cli tests are now part of the main e2e test suite, there is
no need to keep other tests (the cypress part was only browsing the repo
and there is already a test that does that in e2e full suite).

Nightly distlp tests are removed and the associated tools to make the
(hellish) setup of the target infra.

Svnroot push and worker:svnroot-update commands were only used for
distlp and are removed as well.

Fare well distlp, thanks for all the fish !

Closes request #29998 Remove Distributed Tuleap (distlp) support

Change-Id: I9407196cfc19b189729612088d844f4fa7e194f6