feat: Do not be able to modify mirrored milestones links in the artifact UI
This is part of request #30358 Mirrored milestones link cannot be delete in the Tuleap UI

In order to prevent wrong deletion of artifact links that handle the mirror of programs,
we have to prevent users to delete this kind of artifact links in the Web UI.

This patch removes the delete checkboxes for ```_mirrored_milestone```
links and the ability to update the type of these links.

How to test
-----------

* Go to an artifact with at least one artifact link of type ```_mirrored_milestone```
* Switch the artifact link in edit mode

=> There is no checkboxes to delete the link of type
```_mirrored_milestone``` in the Type pane

=> There is no checkboxes to delete the link of type
```_mirrored_milestone``` in the tracker pane and the type selectbox
must be disabled.

Note: Artifact modal will be handled in a dedicated commit.

Change-Id: I73d595e7dacea20d4dce0acd385c7c5349335228