feat: Restrict server
Site admin can restrict and unrestrict the document server.

There is no automagic whitelisting of existing projects because the
feature targets instances where OO cannot be deployed yet because of the
lack of multiple servers. For other cases, a danger modal is displayed
to warn the admin.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I26ac6c1ca01dbc4c0b2b2730db732c37a5c1d667