feat: Reduce nb click when adding/removing projects
Now adding or removing a project submit the form, no need for another
click.

This aligns the behavior with other restrict-resource-access apps in
Tuleap (plugins administration, git mirrors, …).

We need to explain why moving or removing a project is dangerous. Therefore:
* Before adding a project to a server a confirmation modal is displayed
if it was already allowed to use another server.
* Before removing a bunch of projects from a server a confirmation modal
is displayed.

Part of story #29981: restrict an OnlyOffice server to some projects only

Change-Id: I6686b7e154e625901ddbce8cd819c03f76a100b8