feat: add theme variant name to integration route
This can be useful to serve the approriate favicon for instance.

The information is added to 3rd_party_integration_data route even if
it's a user related data because we don't want to mandate 3rd party
applications to issue another REST call to get the user info. Moreover
it's kind consistent with the fact that stylesheet (user dependent) is
already there. It's GraphQL as code sort of :)

Part of: request #28829 Tuleap Skin improvements (as of Tuleap 14.0)

Change-Id: I1aaca8559a675cf4018c826e1acbc1d22bbd2182