fix(el9): htmldoc package is broken on el9
htmldoc package was made on the example of epel's version of the package
that did fiddling on fonts to use the one shipped by official packages.
The package urw-base35-fonts-legacy is no longer shipped in el9,
replaced by urw-fonts but the matching is not that easy and seems
un-necessary for our use case. It's more simple to use the fonts shipped
with upstream htmldoc.

The patch also fix the provides/obsoletes syntax so the dependency can
be correctly resolved.

Part of: request #29619 Add support of Rocky Linux 9

Change-Id: Ic1f7149a52c96ec8bb23f427627aa49cfde16b2f