feat: Dedicated error page when confirmation link cannot be sent
If the delivery of the email containing the confirmation link fails,
then a dedicated error page is displayed to the end user.

This avoids to have the current situation where an error feedback is
displayed next to "Confirmation link sent": user don't know which
message is valid: the feedback or the title of the page.

Part of request #29605: Convert account registration pages to BurningParrot

Change-Id: I8401bb54639165804593aa14c49ac9dbf35f452d