feat: Tuleap backend must blocks deletion of _mirrored_milestones links
This is part of request #30358 Mirrored milestones link cannot be delete in the Tuleap UI

The artifact update must be rejected if at least one artifact link of
type ```_mirrored_milestone``` is deleted.

How to test
-----------

* Go to an artifact with at least one artifact link of type ```_mirrored_milestone```
* Mark this link for deletion and submit the artifact

=> The artifact update must be rejected (in both artifact UI and
artifact modal).

Note: The update of this kind of artifact link (i.e change from
```_mirrored_milestone``` type to another type) will be handled in a
dedicated commit.

Change-Id: Iaf78313aac3ec0fd98bf7deec5adb3b453cfd457