feat: Add token in registration link in invitation email
When a user invites someone to register on Tuleap, the latter should not
have to fill it's email, and will not receive a confirmation link by
email.

Furthermore, when there is no need for user approval by site admin
(`sys_user_approval = 0`) then the user is automagically logged in and
redirected to its dashboard. A success feedback is currently displayed
but it will be replaced by modal in a dedicated contribution.

User approval is left unchanged if `sys_user_approval = 1`. This means
that site admin can be tempted to ask to the invited user to confirm its
email with [Validate] button on Pending users page (And
[Resend email] button on Users pending email activation). A dedicate
contribution will deal with this if needed.

If token is invalid, then request is rejected.

Note: this change of behavior is only valid for new invitations. Old
invitations still works, but invitees will have to confirm their email.

Suggested starting points for the review:
* token is generated in src/common/InviteBuddy/InvitationSender.php
* token is then used to retrieve the invitation in src/common/User/Account/Register/InvitationToEmailRequestExtractor.php
* so that email is forced in src/common/User/Account/Register/RegisterFormPresenterBuilder.php and in src/common/User/Account/Register/RegisterFormHandler.php

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: Ic270cebc45921a87e0090c4ceea6268a51ab5a95