feat: Invited users don't need email validation
When the platform is configured for user approval (can be configured
with `tuleap config-set sys_user_approval 1`) and user registers via an
invitation, then [Validate] button is useless because user email is
already validated with the token in the invitation.

In that case the button is disabled (as well as its counterpart [Resend
email] when platform is switched back to sys_user_approval=0). Backend
do not send unnecessary email if user manage to forge the request (by
manully removing the disabled attribute for example).

Bonus: redirect to self page is added for each actions to prevent form
resubmission.

Technical note: It would have be best to first refactor the
approve_pending_users.php before adding more crap in it. However the
amount of needed refactoring is pretty big and refactoring of
register.php has already sucked all my will :( There will be dedicated
contributions to do it later, if time permits it.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: Ib93e000c402a55fd045f0ccb9beda22edfba0cbf