refactor: Remove $status from create method
create() was accepting a parameter @status, however it was always the
same "creating". No need to pass it as parameter then, it will simplify
the reading.

Furthermore the method update() that was used to update the status of an
invitation to "sent" or "error" is removed in favor of two dedicated
methods that should have more business meaning: markAsSent/markAsError.

No functional changes.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: I678d1468dd8419b96ae113c0835f9c4ba4cb7cbc