fix: wrong redirection for /my/
As anonymous user, try to go to the following url:
https://tuleap-web.tuleap-aio-dev.docker/my/?toto=titi&tata=tutu

Should be asked to login, and then should be redirected to this exact
url, not with a return_to parameter.

Without this, welcome modal is not displayed if user tries to access to
her personal page with an invitation token while she is not logged in.

Honestly, this part of the code is a mystery. I don't really understand
the use cases for a special treatment for "/my/" urls. I tried to insert
myself without touching too much as I am afraid to break something
already brittle.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: I2d704f4ea16b37c37366aec75e5596a003edb756