fix: Close request #30387 Symbolic links for legacy SVN and CVS must be recreated at each container start
In the tuleap-aio-dev Docker container, we have to recreate some stuff in order
to have a working Tuleap instance.

This recreation must be done for the symbolic links for legacy SVN and CVS services.
Without that, the Run Tuleap Dev fails and nginx must be restarted by hand to be working.

How to test
----------

* After applying this patch, restart your tuleap-aio-dev containers

=> All must be well started (nginx ...), /svnroot and /cvsroot symbolic
links must still be created and have all the expected content

Change-Id: I7cadd1075e8708d869d98bdbcb4e2612b19e97b3