feat: Always display a welcome modal
This changes targets the invitees behind a sys_user_approval=1, but will
benefit for every users (without invitation, and/or when
sys_user_approval=0).

On successful login, a first timer is welcomed with a modal.
If the user has been invited by another user, then the modal
displays this user.

Note: the modal is displayed only when user is redirected to /my/ after
successful login.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: I78d9879900d1c69f05767f2aee421866c2787e97