fix: Close request #30392 Selected parent in modal artifact link field lead to an unknown URL
When editing an artifact with the modal, if you select a parent from the possible
parent list provided by the artifact link field, then a preview of this new link appears.
If we open the link in this preview, this leads to a wrong URL /plugins/agiledashboard/undefined.

If we select the new parent by entering directly its ID, it works as expected.

How to test
-----------

* Open an artifact without any parent in edition with the artifact modal
* Go to the artifact link field, and select a possible parent from the
  list provided by the field

=> The link preview has now a correct link to go to the artifact

Change-Id: I561d79f7032852f8663c016132882e0b0402b4a1