feat: Clear email from invitations table whenever possible
We should not keep forever a PII like the email address when the
invitation is flagged as error.

Likewise, we should not duplicate the email address when the invitation
is successfully used, user table will contain the email, no need to have
it at both places.

To test:

* Both Alice and Bob invites charlie@example.com
* Charlie clicks on the invitation link in Bob's email message, and
register, and gets a welcome modal.
* Both Alice and Bob have received an email to warn them about Charlie's
registration
* Charlie logout, and click again on invitation links: a message tells that
the invitation is already completed and charlie should login instead.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: Id7c612416acdf5f1553c9164ff446d619cfe554f