fix: request #30371 Artifact modal should expose errors raised by backend
As of today, when there is a REST error in backend due to validation,
the detail of what failed is not displayed to user.
A generic error message is displayed while the error is there in the REST payload.

How to test
-----------

* Go to a project that have TTM activated with a bug tracker defined in
  TTM configuration
* For this Bug tracker, update the title field in Tracker administration
  to allow 3 chars max
* Go into a campaign, and then into a test
* Ask to create a bug that will be linked to this test with a title that
  has more than 3 chars

=> The creation is not done, and the explicit error message returned by
the REST route must be displayed. The same goes in any other service
that uses the modal.

Change-Id: Ia0d4b9139b0fe07e67a58c823129352d8de59c35