feat: Remove obsolete invitations
Default lifetime of an invitation is now 30 days. After that, pending
invitations are removed. Admins can change the 30 days lifetime to
another value (e.g 15 days) with the following command:
`tuleap config-set invitations_lifetime 15`

Purged invitations will be processed in a dedicated contribution to send
an email to warn the users who invited people in vain.

To test forge your DB with time in the past for some invitations, then
run `tuleap daily-job`. Unprocessed invitations that are obsolete are
removed. Processed invitations are kept.

Part of request #29614: Invited users shouldn't be mandated to re-confirm their email

Change-Id: Ie3c061d5074fbb04f48ad81ecbabee7dfc8371ea