Be able to remove reverse link artifact
Part of story #24970 [modal] select type for links

When an artifact source ID is not provided in the `all_links` key and
if this source artfact was already linked with the updated artifact then
link is removed. If the source artifact ID is provided in the key then
the link between the source aritfact and the updated artifact still
exists.

How to test:
 - Make sure to have the feature flag enabled => $ tuleap config-set feature_flag_allow_reverse_links 1
 - In our example the artifact we want to update is 200. It is already reversed linked to the artifacts 10 and 12.
 - PUT artifacts/200 and the following payload
```
{
  "values": [
    {
      "field_id": 154,
      "all_links": [
        {
          "id": 10,
          "type": "fixed_in",
          "direction": "reverse"
        },
        {
          "id": 45,
          "type": "fixed_in",
          "direction": "reverse"
        }
      ]
    }
  ],
  "comment": {
    "format": "commonmark",
    "body": ""
  }
}
```

Expected results:
 - The link between 200 and 12 is REMOVED
 - The link between 200 and 45 is ADDED
 - The link between 200 and 10 is KEPT

Note:  We cannot remove all the reversed links when `all_links` is equal
to [], it will be done in a future contribution

Change-Id: Icf43eac8f036de3342c963d09ad40b5ac605b0b0