Fix request #30406 Docker image doesn't handle properly switch to PHP 8.1
The work for container startup sequence was done before we moved to the
ubiquitous `tuleap-cfg site-deploy` command and the original issue
(docker image without php 8.1 config) comes from the fact we forgot to
bump to php 8.1 in the docker startup sequence.

It's more efficient and trustworthy to run the `site-deploy` command as
part of the container startup sequence to avoid future discrepancies.

Change-Id: Ia264e6ec609723b4bcad15f0231d822c882edc86