fix: None should not add user invited to project
In project history, the user who add a new project member thanks to an
invitation should not be "None", but the user who invited the project
member.

This was due to the fact that deep down in the call stack,
groupAddHistory was adding the log by requesting the current user from
the request. This does not work in our case because the user is
currently registering herself, so there is no current user.

We should alway pass the project admin who is doing the operation so
groupAddHistory is now marked as deprecated.

Took the opportunity to switch ProjectHistoryDao to EasyDB, there should
not be any functional changes in project history, project history
export, or site admin project change history.

You can test the invitation by updating by hand an ongoing invitation in
the db to add a project in `to_project_id`: once completed, the user is
still added to the project, and the corresponding entry in the project
history is not made by None.

Bonus: InvitationInstrumentation::incrementUsedInvitation() is now used
when user is successfully invited to a project.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Ife92454708b943d0076f72b4ba67b8315a3fcdac