fix request #30419 Incorrect project visibility at project creation
How to test:
- You can throw an exception in ProjectResource to assert that the right
payload is sent.
- When you create a new project and leave the default visibility (for
  example "private", then your project is created as Private, not
Public.
- When you choose another visibility, then your project is created with
it.

Notes:
- the "configuration" key in the root store was getting overwritten.
- the namespaced State annotation was not working and returned
`undefined`. Using the state directly to init the v-model is more stable
and less complicated.
- I've run into a race condition: a parent mounted() is always called
after all its children's mounted(). Since the child dispatches an event
in its mounted() method and the parent listens to that event, the
initial event was lost because the parent did not have time to register
itself as a listener. This was probably a side-effect of list-picker
becoming synchronous: previously, we had to await it, which gave time to
the parent to register itself.

Change-Id: Ibce5b9a2a1d2b71dd4edc4f701eecf27ac44daff