feat: Be able to reopen an abandoned pull request in UI
This is part of request #30013 Be able to reopen an abandoned pull request

When accessing an abandoned pull request in the UI, if the user is able
to reopen it, then a reopen button is now displayed and clickable.

How to test
-----------

* Go to an abandoned pull request where you are able to write in both
  source and destination repositories and branches

=> A "reopen" button is now displayed and functionnal.

* Go to an abandoned pull request where you are not able to write in both
  source and destination repositories and branches

=> The disabled "Abandoned" button is still displayed.

Note: Maybe the reopen right is too restrictive for now (user must be
able to write in both source and destination repository ans branches).
We can still change it.

Change-Id: I4260156be6e0cbfd59d9e80aef77fd31353f9d10