feat: Close request #30372 Do not silently truncate text input
When trying to create a new artifact with text in text field that is larger
than what underlying storage (mysql TEXT atm) accept, the artifact creation
should report an error rather than truncating the input.

How to test
-----------

* Go to an artifact with a text field
* Add text content to this field with more than 65535 chars

=> The creation/update of the artifact must fails with an explicit error
message.

Note 1: The maxlength HTML attribute does not seem to take into account
spaces and new lines (at least in Firefox).

Note 2: CKEditor does not deal natively with maxlength. A dedicated
development must be done for that and I propose to do this in a
dedicated commit.

Change-Id: Ifb3b9d9e95e953b7bb62908da01e4883957ff8ec