feat: Allow project admin to withdraw an invitation
If an invitation has been made for a project, then in project » admin »
members » pending invitations, there is a button to withdraw the
invitation.

All pending invitations for the same email for the same project are
removed.

A log is added in project history (Not sure about the usefulness since
the email is not logged to not store PII).

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: I950f909b83d8240714a5702b20109cd26b6f3702