Drop PHPCompatibility sniffs
They are not as useful as they were initially and are lagging behind the
current PHP versions.
Since the time we introduced them we now have the codebase shape have
vastly changed and we now have static analysis via Psalm to keep us on
our toes.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Id6cebf7f7a4ac0d83dc7de7e5b090e06e747d49d