Be able to update forward links
Part of story #24970 [modal] select type for links

How to test:
  - Enable the feature flag:
    $ tuleap config-set feature_flag_allow_reverse_links 1
  - Make a PUT request to an existing artifact (e.g 120) with the following payload:

```
{
  "values": [{
    "field_id": 543,
    "all_links": [
      {"id": 11, "direction": "reverse", "type": ""},
      {"id": 151, "direction": "forward", "type": "_is_child"}
    ]
  }]
}
```

The artifact 120 should have artifact 11 as reverse link
The artifact 120 should have artifact 151 as forward link

Since the behaviours of the route must not change , some reverse links with particular type such as `_mirrored_links` are ignored. In that case, no error is returned because it would break the current behaviours of other service such as Program Management (see REST test).

Change-Id: I0d688405d38b26764c266fc1eb1549fa1b396f5c