feat: Frontend to invite in a project
In the invite modal, I have now a selectbox that lists the project I am
admin of. The current project, if any, is preselected.

If I select a project, then invitations are made for this project and
users will be automatically part of the project once registered.

If I am not project administrator, then there is no select but I can
still invite buddies.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Icf31132021deeafe5a905cbd216ead162fe826cb