fix request #31058: user preferences e2e tests are failing
It looks like that the failing assertion should have never worked:

```
cy.get("[data-user-has-accessibility-mode]").contains(1);
```

If I understand well, it basically search for the text `1` in the
`body`… 🤔
Now we check that the data attribute is equal to `1`.

Took the opportunity to be able to run multiple times the tests because
it is a real PITA to run user pref tests with the same account, since
changes are not reset after tests. On the bonus side, this adds a basic
coverage of user registration.

Change-Id: Ie41b7a63638cb9e4791dfc50a0736578685c91b9