feat: ability to resend an invitation
In project » admin » members » pending invitations, the project admin is
now able to resend an invitation.

For now the new invitation is sent on behalf of the current user and the
custom message is not kept (invitee should not think that one user said
something that is actually said by another one).

We might want to include both users in the invitation email, but we need
clearer design decision on this subject beforehand. It will be part of a
dedicated changeset if needed.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: I7f2a6549c48859e7c7b632ad00f9d87b11037849