feat: log sent invitation in project history
If a user invites someone in a project, we should add an entry in
project history.

If a pending invitation is resent, then only one event "resent" appears
in project history (instead of resent+sent).

Furthermore, when a user invites someone in a project, we now increment
the project counter for instrumentation instead of the platform one.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Ifd901fc7dd27640e88cecbea44e2dcc2a95c6bb2