fix: Account creation feedback not sent in the right format
When an invitation is complete, the user who invited the new member
receives a confirmation email. This email should be both in text/html
and text/plain alternatives.

Currently the text/plain content is sent as text/html format which is
wrong.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: I91d6aff8037e3bbeab30bb2d139624a5ab913169