fix: Close request #31062 System service can be wronlgy used
When creating a new system service (which is a service not activated in the template
used to create the current project; it only concerns MW standalone as of today),
this service can be wrongly activated in project, even if plugin disallow this activation.

Submitting the edit service modal (with our without any modifications) switches the service usage to "used".

How to test
-----------

* Create a new project with a template that does not have the MW
  standalone activated and has the legacy MW (1.23) activated
* In this new project, switch to used if needed the MW 1.23 service.
* Switch to unused the MW 1.23 service
* Open the service edition modal for MW standalone and submit it without
  any modifications

=> The MW standalone service must not be switched to used.

Change-Id: I44678ffeafd9afdc7774d55538402e7d514b54fc