feat: User with delegation can invite in project
As project admin, in a project create a user group with a project member
in it (make sure that this user is not project admin). Activate
"membership management" delegation for this user group.

As member of this user group, invite someone in the project
(sys_user_approval=0 will make it easier to test). Once registered, the
invitee is added in the project and the history shows that it has been
added by the member.

Those that are not project admin nor have permission delegation cannot
invite buddies in the project.

Took the opportunity to extract the double check admin|delegation in a
dedicated object to not have to repeat ourselves (and make mistakes).

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Iaba0567478f6fce3c24aff605066091df67e8bfc