refactor: extract logic to send one invitation
This reduces responsibility of InvitationSender, and will allow to add
more logic in the new object.

In the future the sending of an invitation might return different things
according to situation (email in error, user added in project, user
already in project, …). To handle separately errors and success, we use
Neverthrow.

No functional changes.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Iacdfbac7aec103f47f7c93649e01df13321c56b0