feat: cover invitation sending by some e2e tests
Add an e2e test to make sure that people with delegated permissions can
send invitation into a project.

This only covers basic invitation management:
* Send invitation
* Resend invitation
* Withdraw invitation

Completing an invitation would be a great addition, but it needs to
parse quoted printable mime parts of an email, not sure this needs to be
part of this contribution (and maybe too brittle to be done at all).

This also changes back the error message when a user tries to access to
administration of a project to fix other e2e tests.

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: I935168ef290761119500d625427efd44d897be3a