feat: Add invitee in project if we already know it
When Alice invites `bob@example.com` in a project and there is already an
existing user Bob with this email then we add Bob in the project and we
don't need to send him an invitation (He will receive a welcome-to
project notification).

If the addition of the user as project member fails, then the reason is
displayed in the feedback modal.

Feedbacks that are added in ManageProjectInvitationsController are here
to cover race conditions: admin resend an invitation and at the same
time the invitation is completed and user added to project. The chances
to get those feedbacks are slim, maybe the default "Invitation has been
resent" would be good enough… 🤔

Part of story #30021: Invite buddies to my Tuleap project

Change-Id: Ibe327dbcfd43a384ecd8f25b17a4337be0aa09d1