Display last CI status badge
part of story #31069 Render everything that is read only

How to test:

In order to test the different states, you can replace the block of
code in PullRequestCIStatus at line 66 by

``` TS
const props = {
    pull_request_info: {
        last_build_status: BUILD_STATUS_PENDING,
        last_build_date: "2023-02-20T10:00:00Z",
    }
}
```

and change the property `last_build_status` with:
- BUILD_STATUS_PENDING
- BUILD_STATUS_SUCCESS
- BUILD_STATUS_FAILED
- BUILD_STATUS_UNKNOWN (set a null last_build_date for this one)

Please note:

As per the mockups,it is expected to have outlined badges for the pending and success states.

It is also expected to have a relative only or an absolute only date display in the badge, since it would take too much space.

For the remaining, it should behave like the CI status badge in the angular app.

Change-Id: I321fd97907476faf825777aaeaf8391b65f0d8a3