Display fetch errors
part of story #31069 Render everything that is read only

This contributions aims to display any error occurring during the
loading of data, like the pull-request or author info.

Since we don't have any mockups regarding the way we should display
those errors, I decided to recreate the same modal as the one we have in
the angular app.

We'll see later with UXD how things should be done, when he'll be
available.

How to test:
- throw a RestException in PullRequestResource line 347
--> The modal opens

- Move the RestException to UserResource line 123
--> The modal opens

In the modal:
- The details are shown when you click [show details]
- The modal is closed when you click [OK] in the footer.

Note: We might want to redirect the user somewhere else, but it is the
current behavior in pullrequest-app, so I guess it is goodenough for
now.

Change-Id: I63f849ecffd18869a378fc12b80705930de89a79